rename HyperGraph to GraphAutoModel extending AutoModel
